feat: classe para métodos estáticos de ordenação e novo endpoint de agendamentos
